# api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "V.SMART API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/currencies/export": {
      "get": {
        "tags": [
          "Currency"
        ],
        "summary": "<b>GET api/v1/currencies/export?format=xlsx</b> — the filtered set as a workbook.\r\n            \r\n             \r\nIt takes the same V.SMART.Api.Contracts.CurrencyQuery the list endpoint does, so the same             query string produces the same rows; only `pageNumber`/`pageSize` are ignored,             because an export is the whole filtered set rather than a page of it.",
        "operationId": "exportCurrencies",
        "parameters": [
          {
            "name": "currName",
            "in": "query",
            "description": "Case-insensitive contains match on the currency name. Wire name `currName`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "createdBy",
            "in": "query",
            "description": "Case-insensitive contains match on the creating user. Wire name `createdBy`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fromDate",
            "in": "query",
            "description": "Inclusive lower bound on `CreatedDate`, applied at the start of the day.\r\n            Wire name `fromDate`.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "description": "Inclusive upper bound on `CreatedDate`, applied at the end of the day.\r\n            Wire name `toDate`.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "description": "1-based page index. Default 1. Wire name `pageNumber`.",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Rows per page. Default 20, maximum 100. Wire name `pageSize`.",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Comma-separated sort fields, `-` prefix for descending — e.g.\r\n`-createdDate,currName`. Omit it to keep the resource's existing default ordering.\r\nValidated against V.SMART.Api.Contracts.PagedQuery.SortableFields; an unknown field is a 400 that lists\r\nthe permitted values. Wire name `sort`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/currencies/import-template": {
      "get": {
        "tags": [
          "Currency"
        ],
        "summary": "<b>GET api/v1/currencies/import-template</b> — the blank workbook a user fills in,\r\n            produced by the unmodified `IExcelTemplateService.CreateTemplateAsync`.",
        "description": "It does <b>not</b> call `DownloadTemplate(uploadType)`: that method's\r\n`LoadHeadings` switch knows four item-oriented upload types only\r\n(`ExcelTemplateService.cs:46-76`) and returns an empty header array for anything\r\nelse, so a \"Currency\" upload type would silently produce an empty template. Passing the\r\nheaders explicitly wraps the same service without modifying it.",
        "operationId": "getCurrencyImportTemplate",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/currencies/import": {
      "post": {
        "tags": [
          "Currency"
        ],
        "summary": "<b>POST api/v1/currencies/import</b> — `multipart/form-data` with `file`.\r\n            \r\n             \r\nEvery row goes through `ICurrencyService.CreateAsync`. The controller parses             the spreadsheet and does nothing else: duplicate detection, validation and persistence             stay in the business service, and its refusal message is reported verbatim             (BR-SO-001). A bad row is reported, not thrown; the remaining rows are still             processed.",
        "operationId": "importCurrencies",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ImportResult": {
        "type": "object",
        "properties": {
          "totalRows": {
            "type": "integer",
            "description": "Data rows found in the sheet, excluding the header.",
            "format": "int32"
          },
          "accepted": {
            "type": "integer",
            "description": "Rows the business service created.",
            "format": "int32"
          },
          "rejected": {
            "type": "integer",
            "description": "Rows refused, each with a reason in Errors.",
            "format": "int32"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportRowError"
            },
            "description": "One entry per rejected row, in sheet order.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The body of a `POST api/v1/{resource}/import` (M2-B06). An import is not a\r\nsuccess/failure: it is a per-row report, because a spreadsheet with one bad row must not\r\ndiscard the other 499 and must not silently accept them either."
      },
      "ImportRowError": {
        "type": "object",
        "properties": {
          "row": {
            "type": "integer",
            "description": "1-based row number as it appears in the spreadsheet, header included.",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "description": "Why the row was refused.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "One rejected row. Message is the business service's own refusal message\r\nverbatim where there is one (BR-SO-001), or the DataAnnotations message from the resource's\r\nViewModel — never a reworded API string."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}